+++ cs/AdventOfCode.2019/Intcode/Models/State.cs
@@ -9,7 +9,7 @@ namespace AdventOfCode2019.Intcode
     public record State
     {
         public bool IsHalted { get; set; }
-        public int MemoryAddress { get; set; }
-        public int RelativeBase { get; set; }
+        public long MemoryAddress { get; set; }
+        public long RelativeBase { get; set; }
     }
 }
